Corrected the ROM version number
The format needs to be:
NNNN.XX  AAYY.VV
NNNN is the name of the ROM. In this case, it is MON3
XX is the target hardware. In this case, it is the 1G
AA is the initials of the Author or Publisher. In this case, it is Brian Chiha, BC
YY is the year of release. In this case it is 24
VV is the version number. In this case, it is 1.5